pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "memex-lab"
version = "0.1.0"
description = "Collector-first research framework for low market cap Solana memecoins."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=12,<16",
    "python-dotenv>=1,<2",
    "base58>=2.1,<3",
]

[project.scripts]
memex-collector = "collector.main:main"

[tool.setuptools.packages.find]
include = ["collector*"]
